app/requirements.txt
flask
flask_cors
GitPython
python-json-logger

